testbackend/requirements.txt
django==3

pymongo[tls,srv]==3.11.2

djangorestframework
django-filter

djoser
djangorestframework_simplejwt
